requirements.txt
Appium-Python-Client>=2.11.1
selenium>=4.11.2
urllib3>=2.0.4 